Move resources to a new package
