+++ systemtests/src/main/java/io/enmasse/systemtest/messaginginfra/ResourceManager.java
@@ -12,6 +12,10 @@ import io.enmasse.systemtest.amqp.AmqpClientFactory;
 import io.enmasse.systemtest.bases.ThrowableRunner;
 import io.enmasse.systemtest.logs.CustomLogger;
 import io.enmasse.systemtest.logs.GlobalLogCollector;
+import io.enmasse.systemtest.messaginginfra.resources.MessagingInfraResourceType;
+import io.enmasse.systemtest.messaginginfra.resources.MessagingTenantResourceType;
+import io.enmasse.systemtest.messaginginfra.resources.NamespaceResourceType;
+import io.enmasse.systemtest.messaginginfra.resources.ResourceType;
 import io.enmasse.systemtest.mqtt.MqttClientFactory;
 import io.enmasse.systemtest.platform.Kubernetes;
 import io.fabric8.kubernetes.api.model.HasMetadata;
