+++ systemtests/src/main/java/io/enmasse/systemtest/messaginginfra/resources/ResourceType.java
@@ -2,7 +2,7 @@
  * Copyright 2020, EnMasse authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.enmasse.systemtest.messaginginfra;
+package io.enmasse.systemtest.messaginginfra.resources;
 
 import io.fabric8.kubernetes.api.model.HasMetadata;
